+++ ThirdPartyAdapters/applovin/applovin/src/main/java/com/applovin/mediation/AppLovinBannerAdListener.java
@@ -13,7 +13,7 @@ import com.applovin.sdk.AppLovinAdLoadListener;
 import com.applovin.sdk.AppLovinSdkUtils;
 import com.google.android.gms.ads.mediation.MediationBannerListener;
 
-/*
+/**
  * The {@link AppLovinBannerAdListener} class is used to forward Banner ad events from the AppLovin
  * SDK to the Google Mobile Ads SDK.
  */
@@ -26,8 +26,8 @@ class AppLovinBannerAdListener
   private final AppLovinAdView mAdView;
   private final String mZoneId;
 
-
-  AppLovinBannerAdListener(String zoneId,
+  AppLovinBannerAdListener(
+      String zoneId,
       AppLovinAdView adView,
       ApplovinAdapter adapter,
       MediationBannerListener mediationBannerListener) {
@@ -40,66 +40,68 @@ class AppLovinBannerAdListener
   // Ad Load Listener.
   @Override
   public void adReceived(final AppLovinAd ad) {
-    ApplovinAdapter
-        .log(DEBUG, "Banner did load ad: " + ad.getAdIdNumber() + " for zone: " + mZoneId);
+    ApplovinAdapter.log(
+        DEBUG, "Banner did load ad: " + ad.getAdIdNumber() + " for zone: " + mZoneId);
 
     mAdView.renderAd(ad);
 
-    AppLovinSdkUtils.runOnUiThread(new Runnable() {
-      @Override
-      public void run() {
-        mMediationBannerListener.onAdLoaded(mAdapter);
-      }
-    });
+    AppLovinSdkUtils.runOnUiThread(
+        new Runnable() {
+          @Override
+          public void run() {
+            mMediationBannerListener.onAdLoaded(mAdapter);
+          }
+        });
   }
 
   @Override
   public void failedToReceiveAd(final int code) {
     ApplovinAdapter.log(ERROR, "Failed to load banner ad with error: " + code);
 
-    AppLovinSdkUtils.runOnUiThread(new Runnable() {
-      @Override
-      public void run() {
-        mMediationBannerListener.onAdFailedToLoad(
-            mAdapter, AppLovinUtils.toAdMobErrorCode(code));
-      }
-    });
+    AppLovinSdkUtils.runOnUiThread(
+        new Runnable() {
+          @Override
+          public void run() {
+            mMediationBannerListener.onAdFailedToLoad(
+                mAdapter, AppLovinUtils.toAdMobErrorCode(code));
+          }
+        });
   }
 
   // Ad Display Listener.
   @Override
   public void adDisplayed(AppLovinAd ad) {
-    ApplovinAdapter.log(DEBUG, "Banner displayed");
+    ApplovinAdapter.log(DEBUG, "Banner displayed.");
   }
 
   @Override
   public void adHidden(AppLovinAd ad) {
-    ApplovinAdapter.log(DEBUG, "Banner dismissed");
+    ApplovinAdapter.log(DEBUG, "Banner dismissed.");
   }
 
   // Ad Click Listener.
   @Override
   public void adClicked(AppLovinAd ad) {
-    ApplovinAdapter.log(DEBUG, "Banner clicked");
+    ApplovinAdapter.log(DEBUG, "Banner clicked.");
     mMediationBannerListener.onAdClicked(mAdapter);
   }
 
   // Ad View Event Listener.
   @Override
   public void adOpenedFullscreen(AppLovinAd ad, AppLovinAdView adView) {
-    ApplovinAdapter.log(DEBUG, "Banner opened fullscreen");
+    ApplovinAdapter.log(DEBUG, "Banner opened fullscreen.");
     mMediationBannerListener.onAdOpened(mAdapter);
   }
 
   @Override
   public void adClosedFullscreen(AppLovinAd ad, AppLovinAdView adView) {
-    ApplovinAdapter.log(DEBUG, "Banner closed fullscreen");
+    ApplovinAdapter.log(DEBUG, "Banner closed fullscreen.");
     mMediationBannerListener.onAdClosed(mAdapter);
   }
 
   @Override
   public void adLeftApplication(AppLovinAd ad, AppLovinAdView adView) {
-    ApplovinAdapter.log(DEBUG, "Banner left application");
+    ApplovinAdapter.log(DEBUG, "Banner left application.");
     mMediationBannerListener.onAdLeftApplication(mAdapter);
   }