+++ ThirdPartyAdapters/inmobi/CHANGELOG.md
@@ -1,7 +1,45 @@
 ## InMobi Android Mediation Adapter Changelog
 
-#### Next Version
+#### Version 9.1.0.0
+- Verified compatibility with InMobi SDK version 9.1.0.
+- Updated the minimum required Google Mobile Ads SDK version to 19.4.0.
+
+Built and test with:
+- Google Mobile Ads SDK version 19.4.0.
+- InMobi SDK version 9.1.0.
+
+#### Version 9.0.9.0
+- Verified compatibility with InMobi SDK version 9.0.9.
+
+Built and test with:
+- Google Mobile Ads SDK version 19.3.0.
+- InMobi SDK version 9.0.9.
+
+#### Version 9.0.8.0
+- Removed support for the deprecated NativeAppInstallAd format. Apps should request unified native ads.
+- Updated the minimum required Google Mobile Ads SDK version to 19.3.0.
+- Verified compatibility with InMobi SDK version 9.0.8.
+
+Built and test with:
+- Google Mobile Ads SDK version 19.3.0.
+- InMobi SDK version 9.0.8.
+
+#### Version 9.0.7.1
+- Fixed an issue where the adapter did not keep InMobi SDK's initialization state properly.
+- Updated the minimum required Google Mobile Ads SDK version to 19.2.0.
+
+Built and test with:
+- Google Mobile Ads SDK version 19.2.0.
+- InMobi SDK version 9.0.7.
+
+#### Version 9.0.7.0
+- Verified compatibility with InMobi SDK version 9.0.7.
+- Updated the adapter to support inline adaptive banner requests.
 - Adapter now includes proguard configuration as suggested by [InMobi's guidelines](https://support.inmobi.com/monetize/android-guidelines).
+
+Built and test with:
+- Google Mobile Ads SDK version 19.1.0.
+- InMobi SDK version 9.0.7.
 
 #### Version 9.0.6.0
 - Verified compatibility with InMobi SDK version 9.0.6.